backend/requirements.txt
fastapi
uvicorn
supabase
pandas
python-dotenv
httpx
google-genai
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib